AdminView breakdown for readability
